moving qml_dialog to it's own file
